+++ qacode/core/webs/controls/control_dropdown.py
@@ -18,10 +18,11 @@ class ControlDropdown(ControlForm):
             Some elements need to search False to be search at future
         """
         kwargs.update({"instance": "ControlDropdown"})
-        strict_rules = kwargs.get("strict_rules")
+        strict_rules = kwargs.get("strict_rules") or []
         if not bool(strict_rules):
             strict_rules.append(
                 {"tag": "select", "type": "tag", "severity": "hight"})
+            kwargs.update({"strict_rules": strict_rules})
         super(ControlDropdown, self).__init__(bot, **kwargs)
         if not self.IS_DROPDOWN and self.tag is not None:
             raise ControlException(msg=MSG.CDD_BADTAG)
@@ -32,8 +33,8 @@ class ControlDropdown(ControlForm):
             if it's neccessary reload element properties
         """
         super(ControlDropdown, self).__check_reload__form__()
-        reload_dropdown_needed = not self.element or not self.dropdown
-        if reload_dropdown_needed:
+        reload_needed = not self.element or not self.dropdown
+        if reload_needed:
             self.reload(**self.settings)
 
     def reload(self, **kwargs):
@@ -42,6 +43,16 @@ class ControlDropdown(ControlForm):
         """
         super(ControlDropdown, self).reload(**kwargs)
         self.dropdown = Select(self.element)
+
+    def __check_dropdown__(self, text, by_value=False, by_index=False):
+        """Internal funcionality for select/deselect methods"""
+        self.__check_reload__form__()
+        if self.dropdown is None:
+            raise ControlException(msg=MSG.CDD_BADTAG)
+        if by_value and by_index:
+            raise ControlException(msg=MSG.CDD_BADPARAMS)
+        if by_index and not isinstance(text, int):
+            raise ControlException(msg=MSG.CDD_BADINDEXTYPE)
 
     def select(self, text, by_value=False, by_index=False):
         """The Select class only works with tags which have select tags.
@@ -63,19 +74,12 @@ class ControlDropdown(ControlForm):
             ControlException -- if tag is not 'select'
             ControlException -- if all flags are 'True'
         """
+        self.__check_dropdown__(
+            text, by_value=by_value, by_index=by_index)
         self.bot.log.debug(MSG.CDD_SELECT_LOADING)
-        self.__check_reload__form__()
-        if self.dropdown is None:
-            raise ControlException(
-                msg="Element must be dropdown, tag={})".format(self.tag))
-        if by_value and by_index:
-            raise ControlException(
-                msg="Can't use this function with all flags with True values")
         if by_value:
             self.dropdown.select_by_value(text)
         elif by_index:
-            if not isinstance(text, int):
-                raise ControlException(msg="index must be an int value")
             self.dropdown.select_by_index(int(text))
         else:
             self.dropdown.select_by_visible_text(text)
@@ -102,15 +106,11 @@ class ControlDropdown(ControlForm):
             ControlException -- if all flags are 'True'
         """
         self.bot.log.debug(MSG.CDD_SELECT_LOADING)
-        self.__check_reload__form__()
-        if by_value and by_index:
-            raise ControlException(
-                msg="Can't use this function with all flags with True values")
+        self.__check_dropdown__(
+            text, by_value=by_value, by_index=by_index)
         if by_value:
             self.dropdown.deselect_by_value(text)
         elif by_index:
-            if not isinstance(text, int):
-                raise ControlException(msg="index must be an int value")
             self.dropdown.deselect_by_index(int(text))
         else:
             self.dropdown.deselect_by_visible_text(text)
@@ -124,6 +124,6 @@ class ControlDropdown(ControlForm):
             ControlException -- if tag is not 'select'
         """
         self.bot.log.debug(MSG.CDD_DESELECTALL_LOADING)
-        self.__check_reload__form__()
+        self.__check_dropdown__('')
         self.dropdown.deselect_all()
         self.bot.log.debug(MSG.CDD_DESELECTALL_LOADED)